Migration to generic intervals
`backup-rotate` can now process arbitrary intervals defined by the user.
An interval is represented by the type Interval, a collection of the
interval prefix, the number of backups to keep for the interval and a
predicate that determines when two backups are considered to be part of
the same timespan for the interval.